pyproject.toml
[project]
name = "multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.1.1",
    "ipykernel>=6.29.5",
    "langchain[google-genai,openai]>=0.3.25",
    "langgraph>=0.4.8",
    "openai>=1.84.0",
    "python-dotenv>=1.1.0",
]
